dprod:: 	seperated cursor position refresh and 	changed in database refresh.
sstrgen_pm::
	changed from types to ones that make more
	sense. (unsigned->size_t)

cfill::
	inital add.
	attempt at making a flood-fill game much
	like the one I've played. At this point
	it just implements the flood-fill algos.
	I have used flood-fill algos but mainly
	the recursive ones. This one uses the
	queue based one & an experimental scan-
	line one (still queue based). Experimental
	as I could not find pseudo code on it.
	Goal is to have the cfill project eventually
	use ncurses to display itself.

+++ srstrgen_pm/ste.cpp
@@ -65,7 +65,7 @@ PartedString::PartedString(){
 PartedString::~PartedString(){
 	delete M;
 }
-unsigned PartedString::lits_size(){
+size_t PartedString::lits_size(){
 	return M->lits.size();
 }
 ostream& operator<<(ostream& sout,PartedString &PS){
